project/requirements.txt
pandas==1.1.4
scikit-learn==0.23.2
psycopg2-binary==2.8.6
fastapi==0.61.2
requests==2.24.0
uvicorn[standard]==0.12.2
gunicorn==20.0.4
orjson==3.4.3